+++ lib/direct_admin.rb
@@ -3,4 +3,8 @@ require_relative "direct_admin/client"
 
 module DirectAdmin
   Error = Class.new(StandardError)
+
+  module Errors
+    InvalidPassword = Class.new(Error)
+  end
 end