+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # master
 - Disable authentication on some commands
+- All requests that return error=1 will raise an exception
 
 # 0.2.0 (2017-04-16)
 - Verify Password command